Refactor tests structure. Remove all not working tests.
