--- app/src/androidTest/java/ru/exwhythat/yather/ui/UiTestSuite.java
@@ -1,13 +1,0 @@
-package ru.exwhythat.yather.ui;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-/**
- * Created by exwhythat on 30.07.17.
- */
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({NavigationTest.class, AboutFragmentTest.class, SettingsFragmentTest.class, WeatherFragmentTest.class})
-public class UiTestSuite {
-}
